refactor(ec2): split 5014-line rest.rs into per-resource submodules
rest.rs was the EC2 long-tail operation sweep at 5014 lines. Convert it from a
single file into a rest/ directory: mod.rs keeps the module doc, re-exports the
shared imports/helpers (region_of, nested_indexed, parse_port_range) and the two
VPC-encryption exclusion consts, and retains the genuinely one-off long-tail
handlers plus the inline test module. Each cohesive resource family moves into
its own submodule, re-exported with 'pub(crate) use <sub>::*' so every
'rest::<op>' dispatch site in service/mod.rs resolves unchanged:

  traffic_mirror (19), route_server (18), capacity_manager (11), fpga (10),
  prefix_list (10), event_window (8), vpc_access (8), vpc_encryption (8),
  byoip (5), ipam (3)

Pure code movement, no behavior change: mod.rs drops to 2329 lines, the 214-fn
set is identical before/after, 123 lib tests pass, workspace clippy -D warnings
clean. (Submodules are one level deeper, so their 'super::tags::tag_set_xml'
refs became 'super::super::tags::tag_set_xml'.)
